Refactor name and package for ProcessManager to NodeExecManager. Actiavte all tests
